% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{align_features_to_standards}
\alias{align_features_to_standards}
\title{Align Features to Standards}
\usage{
align_features_to_standards(
  mzroll_db_con,
  clamr_config,
  peakgroups = NULL,
  mass_spec_standards_con,
  matched_method
)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{peakgroups}{optionally, a subset of peakgroups to generate aligned features from.}

\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}

\item{matched_method}{string representation of configuration details, from standards DB}
}
\description{
For each peakgroup in a dataset identify candidate standard compounds which it could match,
and align experimental and standard features for each putative match.
}
