% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_structure.R
\name{canonicalize_structures}
\alias{canonicalize_structures}
\title{Canonicalize Structures}
\usage{
canonicalize_structures(
  input_vector,
  input_type,
  output_types,
  conda_env,
  conda_path = "auto",
  remove_salts = TRUE
)
}
\arguments{
\item{input_vector}{a character vector of SMILES IDs.}

\item{input_type}{smiles or inchi}

\item{output_types}{e.g., smiles, inchi, inchi-key}

\item{conda_env}{a conda environment where RDKit is installed.}

\item{conda_path}{a valid path to the conda binary or auto to let reticulate
find a default installation}

\item{remove_salts}{TRUE/FALSE whether salts should be removed from a
formula during conversion}
}
\value{
a table mapping the smiles_vector (smiles variable) on
  canonical SMILES
}
\description{
Call a conda installation of RDKit using reticulate to canonicalize and
  interconvert structures.
}
\examples{
\dontrun{
input_vector <- c("O=C(O)C(N)C(O)C", "CC(O)C(N)C(=O)O", "badID")
input_type <- "smiles"
output_types <- "smiles"
conda_env <- "rdkit37"

canonicalize_structures(input_vector, input_type, output_types, conda_env)
}

}
