% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{compare_aligned_features}
\alias{compare_aligned_features}
\title{Compare Aligned Features}
\usage{
compare_aligned_features(
  aligned_labelled_features,
  MS2tol,
  n_top_spectra_summed = 3L,
  quality_weights = c(purity = 2, quality = 1),
  frag_similarity_methods = c("cosine", "robust_cosine"),
  remove_precursor = FALSE,
  minimum_ic_fraction = 0.001,
  print_plot = FALSE
)
}
\arguments{
\item{aligned_labelled_features}{tibble of aligned features with compound identifications.}

\item{MS2tol}{the instrument mass accuracy tolerance for the purpose of fragment grouping - \code{\link[clamr]{format_mass_accuracy_input}}, normally passed from a \code{\link[clamr]{build_clamr_config}} object.}

\item{n_top_spectra_summed}{number of scans to use for consensus spectrum formation.}

\item{quality_weights}{relative importance weight for aspects of scans data (e.g., MS2 purity, quality, etc).}

\item{frag_similarity_methods}{a character vector of methods to use for fragmentation scoring:
\itemize{
  \item{tic: fraction of experimental signal explained by standard fragments.}
  \item{fisher: fisher exact test to look at overlap between experimental and standard fragments.}
  \item{spearman: spearman correlation between experimental and standard fragments.}
  \item{cosine: the cosine similarity between mz-matched ions of the standard and experimental spectra.}
  \item{robust cosine: calculate the cosine similarity when separately dropping out each fragment - then take the worst cosine similarity among them. Identifies cases where a match score is driven by a single fragment.}
 }}

\item{remove_precursor}{Should the precursor mass be removed before matching spectra?}

\item{minimum_ic_fraction}{A value between 0 and 1 indicating for a given peak the fraction of total signal for it to be used for matching.}

\item{print_plot}{option associated with how plot should be printed.}
}
\description{
Compares experimentally derived peaks and scans to standard compound mass, retention time and fragmentation to generate a large set of features.
A subset of these features will inform compound matching.
}
