% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_utils.R
\name{compare_aligned_features_fullscans}
\alias{compare_aligned_features_fullscans}
\title{Compare peakgroup m/z and rt to standard}
\usage{
compare_aligned_features_fullscans(peaks, stdCompounds)
}
\arguments{
\item{peaks}{tibble of peaks information}

\item{stdCompounds}{standard compounds for spectral matching.}
}
\description{
Compare peakgroup m/z and rt to standard
}
