% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_utils.R
\name{compare_aligned_features_peakgroup_qc}
\alias{compare_aligned_features_peakgroup_qc}
\title{Summarizes of peakgroup overall quality}
\usage{
compare_aligned_features_peakgroup_qc(
  aligned_labelled_features,
  missing_peakAreaTop = 1000
)
}
\arguments{
\item{aligned_labelled_features}{tibble of groupId, peaks and scans created by \code{\link{align_features_to_standards}} or \code{link[clamr]{nest_peakgroup_features}}}

\item{missing_peakAreaTop}{missing value ion count cutoff}
}
\value{
summaries of an experimental peakgroup
}
\description{
Summarizes of peakgroup overall quality
}
