% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_injections.R
\name{cosine_sim_scans}
\alias{cosine_sim_scans}
\title{Cosine Similarity Scans}
\usage{
cosine_sim_scans(scan_data)
}
\arguments{
\item{scan_data}{from \code{read_ms_file_mzR}}
}
\value{
a length one numeric
}
\description{
Calculates the median cosine similarity across all pairs of different scans
}
