% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_injections.R
\name{create_consensus_spectra}
\alias{create_consensus_spectra}
\title{Create Consensus Spectra}
\usage{
create_consensus_spectra(
  mz_features,
  grouping_vars = NULL,
  MSXtol,
  ic_min = 1000,
  ion_scan_number_min = 1L,
  plot_type = "none"
)
}
\arguments{
\item{mz_features}{A set of scans for an ion produced by
\code{\link[clamr]{tidy_mzR_extract_mzs}}.}

\item{grouping_vars}{NULL if no grouping or a character vector of variables
to summarize over when collapsing MS2s (all variables need to be in
\code{ms_scan_header} of \code{mz_features}).}

\item{MSXtol}{mass tolerance corresponding to the type of scan being
aggregated. Provided by \link[clamr]{build_clamr_config}}

\item{ic_min}{m/zs which are not present with an average IC of more
\code{ic_min} for any value of \code{grouping_vars} are filtered.}

\item{ion_scan_number_min}{Filter mzs which are present in fewer than
\code{ion_scan_number_min} scans (across all \code{grouping_vars}). This
filter is applied after filtering for \code{ic_min}.}

\item{plot_type}{should a plot be generated; options are: none, print and return.}
}
\value{
a list containing:
\itemize{
  \item{grouping_var_summary - Information about the scans that were
    collapsed over grouping_vars}
  \item{consensus_scans - masses and ic of consensus MS2 scans}
  \item{summary_plot (optionally) - a summary plot will be added if
    \code{plot_type} = "return"}
 }
}
\description{
Aggregate a set of MS2 spectra across a set of grouping variables into
  consensus spectra.
}
\details{
Currently this function only supports a single mass_accuracy_tolerance for
  the purposes collapsing spectra, if you are collapsing MS1 and MS2
  spectra, call this function twice with different tolerance parameters.
}
