% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_import_formatting.R
\name{dbimport_adduct_to_charge}
\alias{dbimport_adduct_to_charge}
\title{Adduct to Charge}
\usage{
dbimport_adduct_to_charge(adductName)
}
\arguments{
\item{adductName}{vector of adducts / precursor types}
}
\value{
a vector of integer charge values
}
\description{
Adduct to Charge
}
