% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_import_formatting.R
\name{dbimport_attributes_list_to_ms_data_list}
\alias{dbimport_attributes_list_to_ms_data_list}
\title{Attributes Liset to MS Data List}
\usage{
dbimport_attributes_list_to_ms_data_list(
  ion_attributes,
  fragmentationData,
  chemicalClass,
  sampleType
)
}
\arguments{
\item{ion_attributes}{a tbl containing .entry_id, compoundName, inchikey [optional], smiles [optional], adductName, mass, precursorMz, mode, precursorCharge, date [optional], path [optional], energy, msLevel, chromatographyMethod [optional], rtObserved [optional]}

\item{fragmentationData}{a tbl containing .entry_id, mz, ic, and optionally label}

\item{chemicalClass}{e.g., lipids, fatty acids, polar}

\item{sampleType}{e.g., "pure standard", "public", "observed", "insilico_no_intensity", "insilico_w_intensity"}
}
\value{
an \code{ms_data_list} used by \code{\link{standard_db_populate_data}}
}
\description{
Attributes Liset to MS Data List
}
