% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_import_formatting.R
\name{dbimport_nce_conversion}
\alias{dbimport_nce_conversion}
\title{NCE conversion}
\usage{
dbimport_nce_conversion(energy_table, to_NCE = TRUE)
}
\arguments{
\item{energy_table}{a tibble containing energy (in units of NCE or eV), mz and charge}

\item{to_NCE}{TRUE for eV -> NCE, FALSE for NCE -> eV}
}
\value{
a numeric vector corresponding to the rows in energy_table with converted energies
}
\description{
Mapping between mass spec energy conventions - between normalized collision energies and electron volts.
}
