% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_import_formatting.R
\name{dbimport_nest_standard_by_key}
\alias{dbimport_nest_standard_by_key}
\title{Nest standards by key}
\usage{
dbimport_nest_standard_by_key(ms_data_list, library_key, chunk_size = NA_real_)
}
\arguments{
\item{ms_data_list}{a list containg mass spectrometry experimental data.
  Three entires are required: ions, samples and measurements:
\itemize{
  \item{ions}
  \item{samples}
  \item{elutions [optional]}
  \item{fragmentations [optional]}
  \item{fragmentationData [optional]}
}}

\item{library_key}{a tibble containing path and library_id}

\item{chunk_size}{NA or numeric chunk size of entries to simultaneously database}
}
\value{
a nested tibble containing library_id and library_data (a list of ms_data_lists)
}
\description{
Join library_key to ms_data_list$standards by path and separate ms_data_list by libraray_id
}
