% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_import_formatting.R
\name{dbimport_parse_collision_energies}
\alias{dbimport_parse_collision_energies}
\title{Parse Collision Energies}
\usage{
dbimport_parse_collision_energies(
  ion_attributes,
  energy_name,
  default_energy_type = NA_character_
)
}
\arguments{
\item{ion_attributes}{a tbl containing .entry_id, compoundName, inchikey [optional], smiles [optional], adductName, mass, precursorMz, mode, precursorCharge, date [optional], path [optional], energy, msLevel, chromatographyMethod [optional], rtObserved [optional]}

\item{energy_name}{Variable name (quoted) specifying collision energies}

\item{default_energy_type}{Default energy type (options are eV, NCE or NA [default])}
}
\value{
a character vector of energies (to allow for stacked collision energies) formatted in NCEs
}
\description{
Determine whether energy is eV or NCE based on annotations and then summarize energies in NCE
}
