% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp.R
\name{download_msp_library}
\alias{download_msp_library}
\title{Download .msp Library}
\usage{
download_msp_library(
  savepath,
  library_url,
  unzip_dl,
  method_dl = "libcurl",
  overwrite = FALSE
)
}
\arguments{
\item{savepath}{Output path}

\item{library_url}{Url of the library to download}

\item{unzip_dl}{True if the file is a .zip file}

\item{method_dl}{method to download libraries from the internet (default: "libcurl")}

\item{overwrite}{Overwrite existing results}
}
\value{
0 invisibly
}
\description{
Download a public msp fragmentation library.
}
