% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_query.R
\name{extract_standard_ions}
\alias{extract_standard_ions}
\title{Extract All Standard Ions}
\usage{
extract_standard_ions(
  standards_data,
  query_ionIds = NULL,
  chromatographic_method = NULL,
  collision_energies = NULL,
  is_lipid = FALSE
)
}
\arguments{
\item{standards_data}{output of \code{\link{query_standards_subset}}.}

\item{query_ionIds}{a vector of ionIds (or all ions if NULL).}

\item{chromatographic_method}{NULL or a length 1 character vector naming a chromatography method to use for retention time querying. Retention times are drawn from any ions and mode of a compound where the chromatographic method matches.}

\item{collision_energies}{NULL or a length 1, comma-separated set of collision energies.}

\item{is_lipid}{flag to describe if this is lipid data in the database, or metabolite data. If it is, the extraction is much simpler and faster.}
}
\value{
a list containing:
\itemize{
  \item{query_ions: ions & their retention time (for the specified chromatographic_method).}
  \item{query_fragmentations: fragmentation measurements that are relevant for the specified ions.}
  \item{query_fragmentationData: measurement data [mz, ic] for the query_fragmentations.}
  }
}
\description{
Extract All Standard Ions
}
\examples{
\dontrun{
ms_mode <- "Negative" # this is available in the polarity field of each scan
chemical_class <- "metabolites"

standards_data <- query_standards_subset(mass_spec_standards_con, chemical_class, ms_mode)
}

}
