% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_query.R
\name{find_standards}
\alias{find_standards}
\title{Find Standards}
\usage{
find_standards(compounds, match_type = "name", mass_spec_standards_con)
}
\arguments{
\item{compounds}{a character vector of compound names or molecular formulas.}

\item{match_type}{type of match to use.
\describe{
  \item{name}{exact case-insensitive substring matching}
  \item{formula}{exact formula match}
}}

\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}
}
\description{
Search for standards by name or formula
}
\examples{
\dontrun{
compound_names <- c("N-acetyl-L-aspart", "Acetyl CoA")
find_standards(compound_names, match_type = "name", mass_spec_standards_con)

compound_formulas <- c("C6H9NO5", "C23H38N7O17P3S")
find_standards(compound_formulas, match_type = "formula", mass_spec_standards_con)
}

}
