% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp.R
\name{format_msp_libraries}
\alias{format_msp_libraries}
\title{Format MSP Libraries}
\usage{
format_msp_libraries(
  msp_compound_attributes,
  comment_function,
  attribute_sanitization_table,
  reduce = FALSE
)
}
\arguments{
\item{msp_compound_attributes}{a list of fragmentation attributes and
data produced by \code{\link{read_msp_feathers}}}

\item{comment_function}{function to use when parsing comments}

\item{attribute_sanitization_table}{lookup tibble between attribute and
new_attribute}

\item{reduce}{if TRUE then sample 100 entries for each file type to test
whether attributes can be reliably matched. If this done then attribute
-> new_attribute mappings will be returned rather than the primary output.}
}
\value{
a tibble of fragmentation attributes for valid entries which have
  a name, mass, precursorMz and adductName
}
\description{
Reformat the output \code{read_msp_feathers} so that attributes use a common
terminology and flag entries which are missing key attributes.
}
