% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_export.R
\name{format_msp_list}
\alias{format_msp_list}
\title{Format MSP List}
\usage{
format_msp_list(db_ions, db_spectra, save_name, dir_path)
}
\arguments{
\item{db_ions}{list of attributes produced by \code{\link{generate_msp}}}

\item{db_spectra}{list of spectra produced by \code{\link{generate_msp}}}

\item{save_name}{save name of msp output file}

\item{dir_path}{path to output folder}
}
\value{
writes one tsv for each energy set
}
\description{
Format MSP List
}
