% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_export.R
\name{format_one_msp_entry}
\alias{format_one_msp_entry}
\title{Format One MSP entry}
\usage{
format_one_msp_entry(one_ion, one_spectra)
}
\arguments{
\item{one_ion}{one ion from \code{\link{format_msp_list}}}

\item{one_spectra}{one ion from \code{\link{format_msp_list}}}
}
\value{
a one column tibble
}
\description{
Format One MSP entry
}
