% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_export.R
\name{generate_msp}
\alias{generate_msp}
\title{Ion Standard Fragmentations}
\usage{
generate_msp(
  nested_ions,
  query_ionIds = NULL,
  normalize_intensities = FALSE,
  minimum_ic_fraction = 0.001,
  save_name,
  dir_path
)
}
\arguments{
\item{nested_ions}{generated by \code{\link{extract_standard_ions}}}

\item{query_ionIds}{subset of ionIds to use (or NULL for all)}

\item{normalize_intensities}{TRUE/FALSE whether to normalize spectra
before summing them}

\item{minimum_ic_fraction}{A value between 0 and 1 indicating for a given peak the fraction of total signal for it to be used for matching.}

\item{save_name}{save name of msp output file}

\item{dir_path}{path to output folder}
}
\value{
writes a set of .msp files for each energy_type to dir_path
}
\description{
Ion Standard Fragmentations
}
