% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_query.R
\name{get_db_create_time}
\alias{get_db_create_time}
\title{Retrieve update string}
\usage{
get_db_create_time(mass_spec_standards_con)
}
\arguments{
\item{mass_spec_standards_con}{connection to cloudSQL database}
}
\description{
Retrieve date and time of last known update to database.
}
\details{
Used as a version, to determine if standards need to be re-extracted, or can safely be re-parsed from saved .rds file.
}
