% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_query.R
\name{get_standards_subset}
\alias{get_standards_subset}
\title{Retrieve Standards Subset}
\usage{
get_standards_subset(
  mass_spec_standards_con,
  matched_method,
  r_scripts_path,
  chemical_class,
  ms_mode,
  chromatographic_method
)
}
\arguments{
\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}

\item{matched_method}{string representation encompassing chemical_class, ms_mode, and chromatographic_method information}

\item{r_scripts_path}{system path containing R scripts, notably mzkit.R}

\item{chemical_class}{Type of chemical of interest: e.g., metabolites}

\item{ms_mode}{Mode mass spec is running in: Negative and Positive}

\item{chromatographic_method}{Chromatography method used}
}
\value{
a list containing relevant entries of the \code{mass_spec_standards} database.
}
\description{
Retrieve a relevant subset of the \code{mass_spec_standards_con} database based on the class of chemicals and mode.
}
\details{
If possible, a pre-computed subset (saved .RDS file) will be retrieved, instead of querying the database directly.
If no applicable .RDS file can be retrieved, the database will be queried directly.
}
