% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{match_id_candidates}
\alias{match_id_candidates}
\title{Match ID Candidates}
\usage{
match_id_candidates(
  mzroll_db_con,
  clamr_config,
  nested_peakgroup_features,
  standards_data
)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{nested_peakgroup_features}{generated from \code{\link[clamr]{nest_peakgroup_features}}.}

\item{standards_data}{generated from \code{\link{query_standards_subset}}.}
}
\value{
id_candidates: a tibble matching each groupId to a set of candidate ionIds
}
\description{
Determine what standards are within MS1 mass tolerance of a set of nested_peakgroup_features
}
