% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_query.R
\name{match_method}
\alias{match_method}
\title{Match Method}
\usage{
match_method(config_list, mass_spec_standards_con)
}
\arguments{
\item{config_list}{a list of named parameters which includes "chromatographic_method", "mode", and "collision_energies" to match to an existing method}

\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}
}
\value{
a matched method string used by \code{\link{get_standards_subset}}
}
\description{
Match Method
}
