% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp.R
\name{parse_all_msps}
\alias{parse_all_msps}
\title{Parse All Msps}
\usage{
parse_all_msps(
  msp_libraries,
  conda_env,
  conda_path,
  path_clamdb_repository,
  big_co = 1
)
}
\arguments{
\item{msp_libraries}{tibble containing
\describe{
  \item{library}{A directory}
  \item{subset}{An individual library}
  \item{filename}{Filename of msps to parse}
  \item{savepath}{Full path of \code{filename}}
}}

\item{conda_env}{a conda environment where RDKit is installed.}

\item{conda_path}{a valid path to the conda binary or auto to let reticulate
find a default installation}

\item{path_clamdb_repository}{path to this repository.
Necessary python scripts are found in subdirectory \code{inst/python},
and will need to be installed and configured for msp parsing to work correctly.}

\item{big_co}{Max .msp file size (in Gb) to automatically process}
}
\value{
0 invisibly
}
\description{
Parse All Msps
}
