% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_query.R
\name{query_standards_subset}
\alias{query_standards_subset}
\title{Query Standards Subset}
\usage{
query_standards_subset(
  mass_spec_standards_con,
  chemical_class,
  ms_mode,
  chromatographic_method,
  sample_filter_quosure = NULL
)
}
\arguments{
\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}

\item{chemical_class}{Type of chemical of interest: e.g., metabolites}

\item{ms_mode}{Mode mass spec is running in: Negative and Positive}

\item{chromatographic_method}{Chromatography method used}

\item{sample_filter_quosure}{An optional quosure that can be used to filter
samples before extracting elutions and spectra}
}
\value{
a list containing relevant entries of the \code{mass_spec_standards} database.
}
\description{
Extract a relevant subset of the \code{mass_spec_standards_con} database based on class of chemicals and mode.
}
