% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_construction.R
\name{query_systematic_compounds}
\alias{query_systematic_compounds}
\title{Query Systematic Compounds}
\usage{
query_systematic_compounds(
  systematic_compounds_con,
  systematicCompoundIds = NULL,
  db_update_version = "recent"
)
}
\arguments{
\item{systematic_compounds_con}{A connection to an SQL database which
stores the structure and systematic IDs of well-studied compounds.}

\item{systematicCompoundIds}{NULL to query all compounds or a numeric
vector of systematic compound IDs}

\item{db_update_version}{multiple versions of the database exist that can be
  differentiated by update date.
\describe{
  \item{all}{Queries all versions of the database. This is useful when
    previously-created systematicCompoundIds need to be matched}
  \item{recent}{Queries the most recent version of the database. This is
    useful when matching to compounds by name or structure.}
  \item{YYYY-MM-DD}{Provide a specific date with the format YYYY-MM-DD
    to query database entries from this date.}
  }}
}
\value{
a list containing:
\describe{
  \item{distinct_compounds}{unique database entries}
  \item{database_ids}{systematic IDs of compounds in KEGG, HMDB, ...}
  \item{compound_aliases}{alternative names for a compound}
  \item{metabolic_pathways}{KEGG pathways associated with a metabolite}
  }
}
\description{
Query Systematic Compounds
}
