% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp.R
\name{run_msp_parse}
\alias{run_msp_parse}
\title{Run MSP parse}
\usage{
run_msp_parse(msp_path, conda_env, conda_path, path_clamdb_repository)
}
\arguments{
\item{msp_path}{path to msp fragmentation file}

\item{conda_env}{a conda environment where RDKit is installed.}

\item{conda_path}{a valid path to the conda binary or auto to let reticulate
find a default installation}

\item{path_clamdb_repository}{path to this repository.
Necessary python scripts are found in subdirectory \code{inst/python},
and will need to be installed and configured for msp parsing to work correctly.}
}
\value{
attributes and fragmentations .feather tables in the
  \code{msp_path} directory
}
\description{
Parse msp file into feather tables using python.
}
