% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_scoring.R
\name{score_meta_svm}
\alias{score_meta_svm}
\title{Score Meta Features using SVM}
\usage{
score_meta_svm(meta_features, svm_model)
}
\arguments{
\item{meta_features}{tibble generated from \code{\link{meta_featurize_scores}}}

\item{svm_model}{parameters of \code{e1071} SVM}
}
\description{
Score Meta Features using SVM
}
