% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_scoring.R
\name{score_standard_matches}
\alias{score_standard_matches}
\title{Score Standard Matches}
\usage{
score_standard_matches(
  comparison_features,
  grouping_vars = "groupId",
  score_method,
  ...
)
}
\arguments{
\item{comparison_features}{Output of \link{compare_aligned_features}.}

\item{grouping_vars}{A character vector of variables which define a unique peak group entry.}

\item{score_method}{A method for scoring the features.}

\item{...}{additional arguments to pass to scoring methods}
}
\value{
comparison_features with \code{score} and \code{is_match} added as variables.
}
\description{
Score Standard Matches
}
