% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_scoring.R
\name{score_standard_matches_baseline}
\alias{score_standard_matches_baseline}
\title{Score Standard Matches Baseline}
\usage{
score_standard_matches_baseline(comparison_features, MS1sd = 1e-06, RTsd = 0.5)
}
\arguments{
\item{comparison_features}{Output of \link{compare_aligned_features}.}

\item{MS1sd}{SD of MS1 tolerance}

\item{RTsd}{SD of retention time tolerance}
}
\description{
Score Standard Matches Baseline
}
