% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_scoring.R
\name{score_standard_matches_baseline_lipids}
\alias{score_standard_matches_baseline_lipids}
\title{Score Standard Matches Lipid Baseline (no RT)}
\usage{
score_standard_matches_baseline_lipids(comparison_features, MS1sd = 1e-06)
}
\arguments{
\item{comparison_features}{Output of \link{compare_aligned_features}.}

\item{MS1sd}{SD of MS1 tolerance}
}
\description{
Score Standard Matches Lipid Baseline (no RT)
}
