% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_scoring.R
\name{score_standard_matches_custom_function}
\alias{score_standard_matches_custom_function}
\title{Score Standard Matches with an arbitrary function and cutoff}
\usage{
score_standard_matches_custom_function(
  comparison_features,
  custom_function,
  cutoff
)
}
\arguments{
\item{comparison_features}{Output of \link{compare_aligned_features}.}

\item{custom_function}{a user-specified function}

\item{cutoff}{cutoff for match FDR control}
}
\description{
Score Standard Matches with an arbitrary function and cutoff
}
