% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_scoring.R
\name{score_standard_matches_lasso}
\alias{score_standard_matches_lasso}
\title{Score Standard Matches Using LASSO Logistic Regression}
\usage{
score_standard_matches_lasso(
  comparison_features,
  lasso_pars_df,
  lasso_co = -Inf,
  return_marginals = FALSE
)
}
\arguments{
\item{comparison_features}{Output of \link{compare_aligned_features}.}

\item{lasso_pars_df}{regression coefficients for lasso logistic regression}

\item{lasso_co}{cutoff for match FDR control}

\item{return_marginals}{TRUE/FALSE whether to add the marginal fit of each covariate}
}
\description{
Score Standard Matches Using LASSO Logistic Regression
}
