% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_scoring.R
\name{score_standard_matches_lasso_svm}
\alias{score_standard_matches_lasso_svm}
\title{Score Standard Matches with LASSO and SVM}
\usage{
score_standard_matches_lasso_svm(
  comparison_features,
  lasso_pars_df,
  svm_model,
  svm_co = 0.5
)
}
\arguments{
\item{comparison_features}{Output of \link{compare_aligned_features}.}

\item{lasso_pars_df}{regression coefficients for lasso logistic regression}

\item{svm_model}{parameters of \code{e1071} SVM}

\item{svm_co}{cutoff for match FDR control}
}
\description{
Score Standard Matches with LASSO and SVM
}
