% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_scoring.R
\name{score_standard_matches_random_forest}
\alias{score_standard_matches_random_forest}
\title{Score Standard Matches with Random Forest}
\usage{
score_standard_matches_random_forest(
  comparison_features,
  forest_model,
  forest_co = 0.5
)
}
\arguments{
\item{comparison_features}{Output of \link{compare_aligned_features}.}

\item{forest_model}{random forest object from \code{randomForest}}

\item{forest_co}{cutoff for match FDR control}
}
\description{
Score Standard Matches with Random Forest
}
