% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_query.R
\name{select_best_fragmentations}
\alias{select_best_fragmentations}
\title{Select Best Fragmentatiions}
\usage{
select_best_fragmentations(target_energies, fragmentations)
}
\arguments{
\item{target_energies}{a length one string of NCE fragmentation energies - if using stacked collision energies then separate energies with ","}

\item{fragmentations}{a set of fragmentations of different provenances with characteristic fragIds}
}
\value{
fragmentations to be used in diverse categories:
\itemize{
  \item{matched_single_energy: match the closest energies to pure standards.}
  \item{all_single_energy: include all pure energies.}
  \item{mixed_energy: match all of the energies of a mixed energy samples [not implemented].}
  \item{insilico_w_intensity: match a theoretical spectra w/ intensities.}
  \item{insilico_no_intensity: match a theoretical spectra w/ only masses.}
  }
}
\description{
Select Best Fragmentatiions
}
