% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_construction_utils.R
\name{sql_class_lookup_table}
\alias{sql_class_lookup_table}
\title{SQL class lookup table}
\usage{
sql_class_lookup_table(dbObj)
}
\arguments{
\item{dbObj}{A object inheriting from \linkS4class{DBIDriver}
or \linkS4class{DBIConnection}}
}
\value{
a named vector mapping from R classes (names) to SQL classses (values).
}
\description{
SQL class lookup table
}
\examples{
dbObj <- RMySQL::MySQL()
}
