% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_construction.R
\name{standard_db_convert_compounds}
\alias{standard_db_convert_compounds}
\title{Convert Compound}
\usage{
standard_db_convert_compounds(compoundId_conversions, mass_spec_standards_con)
}
\arguments{
\item{compoundId_conversions}{tibble containing \code{old_compoundIds} and \code{new_compoundId}}

\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}
}
\value{
update compounds and associated ions in \code{mass_spec_standards_con}
}
\description{
Convert an old compoundId to a new value throughout the database and update ions accordingly
}
