% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_construction.R
\name{standard_db_populate_data}
\alias{standard_db_populate_data}
\title{Populate Standard DB Data}
\usage{
standard_db_populate_data(
  ms_data_list,
  mass_spec_standards_con,
  systematic_compounds_con,
  annotationSource,
  match_var_order = c("inchikey", "name", "alias", "inchikey_noproto",
    "inchikey_connectivity"),
  write_to_db = FALSE,
  strict = TRUE
)
}
\arguments{
\item{ms_data_list}{a list containg mass spectrometry experimental data.
  Three entires are required: ions, samples and measurements:
\itemize{
  \item{ions}
  \item{samples}
  \item{elutions [optional]}
  \item{fragmentations [optional]}
  \item{fragmentationData [optional]}
}}

\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}

\item{systematic_compounds_con}{A connection to an SQL databasing housing
names, structure, aliases and systematic identifiers of known compounds.}

\item{annotationSource}{A unique character string to label the data source.}

\item{match_var_order}{identifier priority order for matching.}

\item{write_to_db}{if FALSE then return the entries which would be
appended to the database, if TRUE then append these to
\code{mass_spec_standards_con}.}

\item{strict}{return warnings in some cases rather than errors.}
}
\value{
a list of tables (if \code{write_to_db} is FALSE) or if
  \code{write_to_db} is TRUE: append new table entries to
  \code{mass_spec_standards_con} as side-effects and return 1.
}
\description{
Add new mass spectrometry standard data present in \code{ms_data_list} to
  the \code{mass_spec_standards_con} database respecting existing entries
  and looking up unique compounds in \code{ms_data_list}.
}
\details{
All supplied tables must include a .entry_id identifier which indicates
  the relationship between tables in the original dataset. This function
  will build mappings between .entry_id and the correponding sampleId,
  compoundId, ionId, elutionId and fragmentationId and append new entries
  to these tables as necessary.

All samples and measurements will be added but only ions and compounds
  which are not already found in the database.
}
\seealso{
\code{\link{standard_db_init}},
  \code{\link{standard_db_populate_support_tables}}
}
