% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_construction.R
\name{standard_db_remove_old_compounds}
\alias{standard_db_remove_old_compounds}
\title{Remove Old Compounds}
\usage{
standard_db_remove_old_compounds(compound_summaries, mass_spec_standards_con)
}
\arguments{
\item{compound_summaries}{output of \code{\link{standard_db_summarize_compounds}}}

\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}
}
\value{
delete compounds from mass_spec_standards_db
}
\description{
Compounds tables are updated as the standard database is populated. Whenever a compound is detected remove the
previous compound summary in case the entry is updated.
}
