% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_construction.R
\name{standard_db_summarize_identifiable_compounds}
\alias{standard_db_summarize_identifiable_compounds}
\title{Standard DB Summarize Identifiable Compounds}
\usage{
standard_db_summarize_identifiable_compounds(
  identifiable_ions,
  mass_spec_standards_con,
  systematic_compounds_con,
  match_var_order,
  db_compound_vars
)
}
\arguments{
\item{identifiable_ions}{filtered ion table from \code{ms_data_list}.}

\item{mass_spec_standards_con}{A connection to an SQL database which stores
experimental mass spectrometry data.}

\item{systematic_compounds_con}{A connection to an SQL databasing housing
names, structure, aliases and systematic identifiers of known compounds.}

\item{match_var_order}{identifier priority order for matching.}

\item{db_compound_vars}{list of parameters based on state of currently processed standards and database.}
}
\description{
Match ions to systematic database, then past standards to determine which
  unique compounds are present. Then determine which to add, or overwrite
  past info.
}
