% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_db_export.R
\name{sum_spectra_by_class}
\alias{sum_spectra_by_class}
\title{Sum Spectra by Class}
\usage{
sum_spectra_by_class(
  stdFragmentation,
  normalize_intensities = FALSE,
  minimum_ic_fraction = 0.001
)
}
\arguments{
\item{stdFragmentation}{subsets of \code{nested_ions}}

\item{normalize_intensities}{TRUE/FALSE whether to normalize spectra
before summing them}

\item{minimum_ic_fraction}{A value between 0 and 1 indicating for a given peak the fraction of total signal for it to be used for matching.}
}
\description{
Sum Spectra by Class
}
