% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_injections.R
\name{summarize_loop_injection}
\alias{summarize_loop_injection}
\title{Summarize Loop Injection}
\usage{
summarize_loop_injection(
  ms_file_path,
  mz_compound,
  minTIC = 0,
  clamr_config,
  ...
)
}
\arguments{
\item{ms_file_path}{full path to MS sample file (mzML or mzXML).}

\item{mz_compound}{standard information.}

\item{minTIC}{Minimum MS2 TIC to consider for filtering MS2 scans.}

\item{clamr_config}{configuration file for clamr search.}

\item{...}{additional arguments to pass to
\code{\link{create_consensus_spectra}}}
}
\description{
This function wraps \code{clamr::tidy_mzR_extract_mzs} and
  \code{\link{create_consensus_spectra}} to generate a summary of MS2
  spectra at various collision energies.
}
