% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_collision_energy}
\alias{test_collision_energy}
\title{Test Collision Energy}
\usage{
test_collision_energy(energy_vector)
}
\arguments{
\item{energy_vector}{a character vector of collision of the form "20,40,80"}
}
\description{
Test Collision Energy
}
